Refactored Vocabulary Models
Grouped together similar models
