+++ filip/semantics/vocabulary/__init__.py
@@ -1,15 +1,28 @@
-from .entity import Entity
-from .class_ import Class
-from .individual import Individual
-from .target_statment import TargetStatement,StatementType
-from .relation import RestrictionType, Relation
-from .data_property import DataProperty
-from .object_property import ObjectProperty
-from .datatype import Datatype,DatatypeType
-from .combined_relation import CombinedRelation
-from .combined_object_relation import CombinedObjectRelation
-from .combined_data_relation import CombinedDataRelation
-from .source import Source
+from .entities import \
+    Entity, \
+    Class, \
+    Individual, \
+    DataProperty, \
+    ObjectProperty,\
+    DataFieldType, \
+    Datatype, \
+    DatatypeType
+from .relation import \
+    TargetStatement,\
+    StatementType, \
+    RestrictionType, \
+    Relation
+from .combined_relations import \
+    CombinedRelation, \
+    CombinedObjectRelation, \
+    CombinedDataRelation, \
+    CDRType
+from .source import Source, DependencyStatement
 
-from .vocabulary import Vocabulary, IdType, LoggingLevel, ParsingError
+from .vocabulary import \
+    Vocabulary, \
+    IdType, \
+    LoggingLevel, \
+    ParsingError, \
+    VocabularySettings
