+++ tests/semantic/test_semantic_models.py
@@ -10,8 +10,7 @@ from tests.config import settings
 from filip.clients.ngsi_v2 import ContextBrokerClient, IoTAClient
 from filip.semantics.semantic_models import SemanticClass, InstanceHeader, \
     Command, DeviceAttribute, DeviceAttributeType
-from filip.semantics.vocabulary.data_property import DataFieldType
-from filip.semantics.vocabulary.vocabulary import VocabularySettings
+from filip.semantics.vocabulary import DataFieldType, VocabularySettings
 from filip.semantics.vocabulary_configurator import VocabularyConfigurator
 from filip.utils.cleanup import clear_all
